+++ pytket/docs/changelog.rst
@@ -10,6 +10,7 @@ Features:
   transform.
 * Add `tk2_to_cx` parameter to `RebaseToCliffordSingles` transform.
 * Reduce depth of circuits resulting from `FullPeepholeOptimise` in some cases.
+* Add `MultiBitOp.multiplier` property.
 
 Fixes: